chore: move draft pages
